Allow LiteVFS to be build as loadable and as linkable extensions
LiteVFS needs to be loaded at runtime for production envs, so we can
use it with any SQLite library.
But in order to run SQLite tests, we need the test driver to be compiled
with LiteVFS.

So this patch adds a `linkable` feature, which switches LiteVFS from
loadable to linkable extension. One `.so` file can't be both, though.